# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/population": {
            "get": {
                "tags": [
                    "Population"
                ],
                "summary": "Get population information.",
                "description": "Return data.",
                "operationId": "getPopulation",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "id_nation": {
                                                "type": "string"
                                            },
                                            "nation": {
                                                "type": "integer"
                                            },
                                            "id_year": {
                                                "type": "string"
                                            },
                                            "year": {
                                                "type": "string"
                                            },
                                            "population": {
                                                "type": "integer"
                                            },
                                            "slug_nation": {
                                                "type": "string"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    }
}